corrigindo questoes 5, 6, 7 e 8 de fundamentos

+++ Ficha fundamentos/Q5-Soma e media de 5 inteiros/q5.c
@@ -22,7 +22,7 @@ int main() {
 
     f = a + b + c + d + e;
 
-    g = (a + b + c + d + e)/5;
+    g = (a + b + c + d + e)/(float)5;
 
     printf("A soma dos valores e: %d", f);
     printf("A media dos valores e: %f", g);
