+++ Ficha fundamentos/Q6- Soma e media de 5 reais/q6.c
@@ -21,7 +21,7 @@ int main() {
 
     f = a + b + c + d + e;
 
-    g = (a + b + c + d + e)/5;
+    g = (a + b + c + d + e)/(float)5;
 
     printf("A soma dos valores e: %f", f);
     printf("A media dos valores e: %f", g);
